+++ phpmyfaq/src/phpMyFAQ/Setup/Update.php
@@ -128,6 +128,9 @@ class Update extends Setup
         $this->applyUpdates400Alpha3();
         $this->applyUpdates400Beta2();
 
+        // 4.1 updates
+        $this->applyUpdates410Alpha();
+
         // Optimize the tables
         $this->optimizeTables();
 
@@ -820,6 +823,54 @@ class Update extends Setup
                     Database::getTablePrefix()
                 );
             }
+        }
+    }
+
+    private function applyUpdates410Alpha(): void
+    {
+        if (version_compare($this->version, '4.1.0-alpha', '<')) {
+            $text = <<<EOT
+User-agent: Amazonbot
+User-agent: anthropic-ai
+User-agent: Applebot-Extended
+User-agent: Bytespider
+User-agent: CCBot
+User-agent: ChatGPT-User
+User-agent: ClaudeBot
+User-agent: Claude-Web
+User-agent: cohere-ai
+User-agent: Diffbot
+User-agent: FacebookBot
+User-agent: facebookexternalhit
+User-agent: FriendlyCrawler
+User-agent: Google-Extended
+User-agent: GoogleOther
+User-agent: GoogleOther-Image
+User-agent: GoogleOther-Video
+User-agent: GPTBot
+User-agent: ICC-Crawler
+User-agent: ImagesiftBot
+User-agent: img2dataset
+User-agent: Meta-ExternalAgent
+User-agent: OAI-SearchBot
+User-agent: omgili
+User-agent: omgilibot
+User-agent: PerplexityBot
+User-agent: PetalBot
+User-agent: Scrapy
+User-agent: Timpibot
+User-agent: VelenPublicWebCrawler
+User-agent: YouBot
+User-agent: Meta-ExternalFetcher
+User-agent: Applebot
+Disallow: /
+
+User-agent: *
+Disallow: /admin/
+
+Sitemap: /sitemap.xml
+EOT;
+            $this->configuration->add('seo.contentRobotsText', $text);
         }
     }